refactor: rename logger file
