+++ src/scan.ts
@@ -1,7 +1,7 @@
 import { env, exit } from 'node:process';
 import kleur from 'kleur';
 import type { Context } from 'types';
-import { log } from './logger';
+import { log } from './log';
 import { stopwatch } from './utils';
 
 function create(ctx: Context) {
